refactor(pp)!: rename pretty_print module to pp
Drop the package-root re-exports of pretty_print symbols. The module is
now imported via `from nclutils import pp` and called as `pp.info(...)`,
`pp.success(...)`, etc., matching the convention used by every other
nclutils submodule. Individual symbols can also be pulled directly with
`from nclutils.pp import info`.

The new shape removes the only exception to the "import from submodule"
rule, makes call sites self-documenting (`pp.info` vs bare `info`), and
eliminates name-collision risk with stdlib `logging` and local
variables.

Renames `src/nclutils/pretty_print/` to `src/nclutils/pp/`,
`tests/pretty_print/` to `tests/pp/`, and `docs/pretty_print.md` to
`docs/pp.md`. Internal stdlib logger names change from
`nclutils.pretty_print.*` to `nclutils.pp.*`.

BREAKING CHANGE: `from nclutils import info, success, ...` and similar
root-level imports no longer work. Migrate to `from nclutils import pp`
with `pp.info(...)` call sites, or `from nclutils.pp import info` for
direct symbol imports.
