--- tests/pretty_print/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for the pretty_print module."""
